requirements.txt
numpy==1.14.5
PyOpenGL==3.1.0
